move MarshallableSpec to another package to specifically list all imports
